# api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "wwg API",
    "description": "The Wasatch Wargamers Campaign App API.",
    "version": "v1"
  },
  "paths": {
...
    "/api/auth/masquerade/end": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Ends a masquerade (decision 0012): back to the Admin's own session, without a password, if\nthe masquerade may still go on. 409 if the session isn't a masquerade.",
        "operationId": "EndMasquerade",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/campaigns/{id}/units": {
      "get": {
        "tags": [
          "ArmyUnits"
        ],
        "summary": "Every unit in the campaign, sorted by army then name (every member): what the map draws\nwhere it may (positions follow the visibility rule).",
        "operationId": "ListCampaignUnits",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArmyUnitResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/armies/{id}/units": {
      "post": {
        "tags": [
          "ArmyUnits"
        ],
        "summary": "Adds library units to the army (Umpire or Admin), each as the campaign's own copy. Only\nunits from the army's factions (409), and none already in the campaign (409).",
        "operationId": "AddArmyUnits",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddArmyUnitsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArmyUnitResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/army-units/{id}": {
      "put": {
        "tags": [
          "ArmyUnits"
        ],
        "summary": "Changes the campaign's copy of a unit: name, type, Fighting Factor and points (Umpire or\nAdmin). The library unit stays as it is.",
        "operationId": "UpdateArmyUnit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArmyUnitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArmyUnitResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "ArmyUnits"
        ],
        "summary": "Removes a unit from the army (Umpire or Admin), while the campaign is setting up; the\nlibrary unit stays.",
        "operationId": "DeleteArmyUnit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/factions/{id}/units": {
      "post": {
        "tags": [
          "Library"
        ],
        "summary": "Adds a unit to a faction (Manager or Admin).",
        "operationId": "CreateUnit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveUnitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/units/{id}": {
      "put": {
        "tags": [
          "Library"
        ],
        "summary": "Changes a library unit (Manager or Admin). Campaigns it's already in keep their copies.",
        "operationId": "UpdateUnit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveUnitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Library"
        ],
        "summary": "Deletes a library unit (Manager or Admin). 409 while it's in a campaign.",
        "operationId": "DeleteUnit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AddArmyUnitsRequest": {
        "required": [
          "unitIds"
        ],
        "type": "object",
        "properties": {
          "unitIds": {
            "maxItems": 100,
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The library units: none already in the campaign."
          }
        },
        "description": "Adds library units to an army, from the factions it takes units from."
      },
...
      "ArmyUnitResponse": {
        "required": [
          "id",
          "armyId",
          "unitId",
          "factionId",
          "nation",
          "name",
          "type",
          "fightingFactor",
          "points"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unit's ID.",
            "format": "uuid"
          },
          "armyId": {
            "type": "string",
            "description": "The army it's in.",
            "format": "uuid"
          },
          "unitId": {
            "type": "string",
            "description": "The library unit it was copied from.",
            "format": "uuid"
          },
          "factionId": {
            "type": "string",
            "description": "That library unit's faction.",
            "format": "uuid"
          },
          "nation": {
            "description": "The nation it marches as (step 45): its faction's, or its army's when the faction has none.",
            "$ref": "#/components/schemas/Nation"
          },
          "name": {
            "type": "string",
            "description": "Its name."
          },
          "type": {
            "description": "What kind of troops it is.",
            "$ref": "#/components/schemas/UnitType"
          },
          "fightingFactor": {
            "type": "integer",
            "description": "Its Fighting Factor (\"FF\"), 1–9.",
            "format": "int32"
          },
          "points": {
            "type": "integer",
            "description": "What it's worth, 0–100.",
            "format": "int32"
          }
        },
        "description": "A unit in an army: the campaign's copy of a library unit."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Nation": {
        "enum": [
          "None",
          "France",
          "Britain",
          "Prussia",
          "Austria",
          "Russia",
          "Spain",
          "Portugal",
          "Sweden",
          "Denmark",
          "Holland",
          "Bavaria",
          "Saxony",
          "Wurttemberg",
          "Westphalia",
          "Baden",
          "Warsaw",
          "Italy",
          "Naples",
          "Brunswick",
          "Hanover",
          "Ottoman"
        ],
        "description": "The nation or state of the Napoleonic wars an army fights for. The front-end draws its flag (a\nsimplified one) as the army's icon; None is a plain flag in the army's colour. Stored by name."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveUnitRequest": {
        "required": [
          "name",
          "type",
          "fightingFactor",
          "points"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string",
            "description": "Its name (duplicates are allowed)."
          },
          "type": {
            "description": "What kind of troops it is.",
            "$ref": "#/components/schemas/UnitType"
          },
          "fightingFactor": {
            "maximum": 9,
            "minimum": 1,
            "type": "integer",
            "description": "Its Fighting Factor (FF).",
            "format": "int32"
          },
          "points": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer",
            "description": "What it's worth.",
            "format": "int32"
          }
        },
        "description": "Adds a unit to a faction in the library, or changes one."
      },
...
      "TokenResponse": {
        "required": [
          "accessToken",
          "expiresIn"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "The bearer token for API calls."
          },
          "expiresIn": {
            "type": "integer",
            "description": "Seconds until the access token expires; refresh before then.",
            "format": "int32"
          }
        },
        "description": "A signed-in session. The access token goes in `Authorization: Bearer …`; the refresh\ntoken is never in a body: it's set as an HttpOnly cookie that only /api/auth receives."
      },
...
      "UnitResponse": {
        "required": [
          "id",
          "factionId",
          "name",
          "type",
          "fightingFactor",
          "points"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unit's ID.",
            "format": "uuid"
          },
          "factionId": {
            "type": "string",
            "description": "Its faction.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Its name."
          },
          "type": {
            "description": "What kind of troops it is.",
            "$ref": "#/components/schemas/UnitType"
          },
          "fightingFactor": {
            "type": "integer",
            "description": "Its Fighting Factor (FF).",
            "format": "int32"
          },
          "points": {
            "type": "integer",
            "description": "What it's worth.",
            "format": "int32"
          }
        },
        "description": "A unit in the library."
      },
      "UnitType": {
        "enum": [
          "LineInfantry",
          "FootArtillery",
          "Engineers",
          "LightInfantry",
          "Partisans",
          "LightCavalry",
          "Scouts",
          "MediumCavalry",
          "HeavyCavalry",
          "HorseArtillery",
          "SupplyTrain",
          "SiegeArtillery",
          "Boat"
        ],
        "description": "What kind of troops a unit is (stored as its name), grouped by the rule book's movement\nclasses (decision 0014): infantry and foot artillery; light infantry and partisans; light\ncavalry and scouts; medium and heavy cavalry and horse artillery; supply and siege artillery;\nand boats, which keep to waterways and lakes (decision 0016)."
      },
...
      "UpdateArmyUnitRequest": {
        "required": [
          "name",
          "type",
          "fightingFactor",
          "points"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string",
            "description": "The unit's name."
          },
          "type": {
            "description": "What kind of troops it is.",
            "$ref": "#/components/schemas/UnitType"
          },
          "fightingFactor": {
            "maximum": 9,
            "minimum": 1,
            "type": "integer",
            "description": "Its Fighting Factor (\"FF\"), 1–9.",
            "format": "int32"
          },
          "points": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer",
            "description": "What it's worth, 0–100.",
            "format": "int32"
          }
        },
        "description": "Changes the campaign's copy of a unit (the library's stays as it is)."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "The accessToken from register, login or refresh.",
        "scheme": "bearer"
      }
    }
  },
...
}